analytics/requirements.txt
asgiref==3.4.1
attrs==21.2.0
bson==0.5.10
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
flask-expects-json==1.6.0
future==0.18.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pymongo==4.2.0
pyparsing==3.0.1
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
yara-python==4.1.3